重构了相机和PLC模型，移除了对 `ObservableObject` 的依赖，简化了实现。增加了多相机和PLC的初始化、资源释放、硬触发模式和消费者线程的支持。更新了UI，增加了显示状态信息的文本块和进度条。移除了不再需要的文件，增加了对 `Microsoft.Net.Compilers.Toolset` 的引用。新增 `MainViewState` 类用于存放UI状态。
Merge branch 'master' of https://github.com/leaf5740/Wpf_RunVision

+++ ViewModels/MainViewModel.cs
@@ -11,6 +11,7 @@ using System.Windows;
 using System.Windows.Input;
 using VM.Core;
 using VMControls.WPF.Release;
+using Wpf_RunVision.Models;
 using Wpf_RunVision.Services;
 using Wpf_RunVision.Utils;
 using Wpf_RunVision.Views;
@@ -30,6 +31,8 @@ namespace Wpf_RunVision.ViewModels
         /// 菜单绑定的方案集合（含"无方案"占位）
         /// </summary>
         public ObservableCollection<string> Schemes { get; } = new ObservableCollection<string>();
+
+        
 
         private VisionCoreService visionCoreService;
 
@@ -213,14 +216,12 @@ namespace Wpf_RunVision.ViewModels
         {
             try
             {
-                // 从工程师切换到员工：直接切换
                 if (CurrentPermission == "当前权限：工程师")
                 {
                     CurrentPermission = "当前权限：员工";
                     CanEditScheme = false;
                     MyLogger.Info("用户权限切换为：员工");
                 }
-                // 从员工切换到工程师：验证密码
                 else
                 {
                     var mainWindow = Application.Current.MainWindow;
@@ -320,7 +321,7 @@ namespace Wpf_RunVision.ViewModels
                     // 更新UI状态
                     CurrentScheme = $"当前方案：{schemeName}";
                     MyLogger.Info($"方案[{schemeName}]加载完成");
-                     visionCoreService = new VisionCoreService(ProjectConfigHelper.Instance.CurrentConfigs.Cameras);
+                    visionCoreService = new VisionCoreService(ProjectConfigHelper.Instance.CurrentConfigs);
 
 
                 });
@@ -340,7 +341,10 @@ namespace Wpf_RunVision.ViewModels
         /// </summary>
         public ICommand WindowClosedCommand => new RelayCommand(() =>
         {
-            visionCoreService.DestroyAllCameras();
+            if (visionCoreService != null)
+            {
+                visionCoreService.DestroyAllCameras();
+            }
             MyLogger.Info("程序正常关闭，所有资源已释放");
         });
     }
